Split large test class in half
