+++ src/databaseTests/src/test/kotlin/org/vaccineimpact/api/databaseTests/tests/burdenEstimateRepository/PopulateBurdenEstimateSetTests.kt
@@ -1,0 +1,163 @@
+package org.vaccineimpact.api.databaseTests.tests
+
+import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatThrownBy
+import org.jooq.Record
+import org.junit.Test
+import org.vaccineimpact.api.app.errors.DatabaseContentsError
+import org.vaccineimpact.api.app.errors.InconsistentDataError
+import org.vaccineimpact.api.app.errors.OperationNotAllowedError
+import org.vaccineimpact.api.app.errors.UnknownObjectError
+import org.vaccineimpact.api.db.JooqContext
+import org.vaccineimpact.api.db.Tables.BURDEN_ESTIMATE
+import org.vaccineimpact.api.db.Tables.BURDEN_OUTCOME
+import org.vaccineimpact.api.db.direct.addBurdenEstimateSet
+import org.vaccineimpact.api.db.fromJoinPath
+import org.vaccineimpact.api.db.toDecimal
+import org.vaccineimpact.api.models.BurdenEstimate
+import org.vaccineimpact.api.models.BurdenEstimateSetType
+import org.vaccineimpact.api.models.BurdenEstimateSetTypeCode
+import org.vaccineimpact.api.models.CreateBurdenEstimateSet
+import java.math.BigDecimal
+
+class PopulateBurdenEstimateSetTests : BurdenEstimateRepositoryTests()
+{
+    @Test
+    fun `can populate burden estimate set`()
+    {
+        var setId: Int? = null
+        var returnedIds: ReturnedIds? = null
+
+        given { db ->
+            returnedIds = setupDatabase(db)
+        } makeTheseChanges { repo ->
+            setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, defaultProperties, username, timestamp)
+            repo.populateBurdenEstimateSet(setId!!, groupId, touchstoneId, scenarioId, data)
+        } andCheckDatabase { db ->
+            checkBurdenEstimates(db, setId!!)
+            checkBurdenEstimateSetMetadata(db, setId!!, returnedIds!!, "complete")
+        }
+    }
+
+
+    @Test
+    fun `cannot populate a set unless status is empty`()
+    {
+        JooqContext().use {
+            val returnedIds = setupDatabase(it)
+            val setId = it.addBurdenEstimateSet(returnedIds.responsibility, returnedIds.modelVersion!!, username, "complete")
+
+            val repo = makeRepository(it)
+            assertThatThrownBy {
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, data)
+            }.isInstanceOf(OperationNotAllowedError::class.java)
+                    .hasMessage("the following problems occurred:\nThis burden estimate set already contains estimates." +
+                            " You must create a new set if you want to upload any new estimates.")
+        }
+    }
+
+
+    @Test
+    fun `populate set throws unknown object error if set does not exist`()
+    {
+        JooqContext().use {
+            setupDatabase(it)
+            val repo = makeRepository(it)
+            assertThatThrownBy {
+                repo.populateBurdenEstimateSet(12, groupId, touchstoneId, scenarioId, data)
+            }.isInstanceOf(UnknownObjectError::class.java)
+        }
+    }
+
+    @Test
+    fun `cannot create burden estimates with diseases that do not match scenario`()
+    {
+        val badData = data.map { it.copy(disease = "YF") }
+        JooqContext().use { db ->
+            setupDatabase(db)
+            val repo = makeRepository(db)
+            assertThatThrownBy {
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, defaultProperties, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, badData)
+            }.isInstanceOf(InconsistentDataError::class.java)
+        }
+    }
+
+    @Test
+    fun `cannot populate burden estimates with non-existent countries`()
+    {
+        val badData = data.map { it.copy(country = "FAKE") }
+        JooqContext().use { db ->
+            setupDatabase(db)
+            val repo = makeRepository(db)
+            assertThatThrownBy {
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, defaultProperties, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, badData)
+            }.isInstanceOf(UnknownObjectError::class.java).matches {
+                (it as UnknownObjectError).typeName == "country"
+            }
+        }
+    }
+
+    @Test
+    fun `cannot populate burden estimate set if cohort_size is missing from burden_outcome table`()
+    {
+        JooqContext().use { db ->
+            setupDatabase(db)
+            db.dsl.deleteFrom(BURDEN_OUTCOME).where(BURDEN_OUTCOME.CODE.eq("cohort_size")).execute()
+            val repo = makeRepository(db)
+            assertThatThrownBy {
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, defaultProperties, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, data)
+            }.isInstanceOf(DatabaseContentsError::class.java)
+        }
+    }
+
+    private fun checkBurdenEstimates(db: JooqContext, setId: Int)
+    {
+        // We order the rows coming back so they are in a guaranteed order. This allows
+        // us to write simple hardcoded expectations.
+        val t = BURDEN_ESTIMATE
+        val records = db.dsl.select(t.BURDEN_ESTIMATE_SET, t.COUNTRY, t.YEAR, t.AGE, t.VALUE, t.STOCHASTIC)
+                .select(BURDEN_OUTCOME.CODE)
+                .fromJoinPath(BURDEN_ESTIMATE, BURDEN_OUTCOME)
+                .orderBy(BURDEN_ESTIMATE.COUNTRY, BURDEN_OUTCOME.CODE)
+                .fetch()
+
+        checkRecord(records[0], setId, 2000, 50, "AFG", "cases", 100.toDecimal())
+        checkRecord(records[1], setId, 2000, 50, "AFG", "cohort_size", 1000.toDecimal())
+        checkRecord(records[2], setId, 2000, 50, "AFG", "deaths", 10.toDecimal())
+        checkRecord(records[3], setId, 1980, 30, "AGO", "cohort_size", 2000.toDecimal())
+        checkRecord(records[4], setId, 1980, 30, "AGO", "dalys", 73.6.toDecimal())
+        checkRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20.toDecimal())
+    }
+
+    private fun checkRecord(record: Record, setId: Int,
+                            year: Int, age: Int, country: String, outcomeCode: String, outcomeValue: BigDecimal)
+    {
+        val t = BURDEN_ESTIMATE
+        assertThat(record[t.BURDEN_ESTIMATE_SET]).isEqualTo(setId)
+        assertThat(record[t.COUNTRY]).isEqualTo(country)
+        assertThat(record[t.YEAR]).isEqualTo(year)
+        assertThat(record[t.AGE]).isEqualTo(age)
+        assertThat(record[t.STOCHASTIC]).isFalse()
+        assertThat(record[BURDEN_OUTCOME.CODE]).isEqualTo(outcomeCode)
+        assertThat(record[t.VALUE]).isEqualTo(outcomeValue)
+    }
+
+    private val data = sequenceOf(
+            BurdenEstimate("Hib3", 2000, 50, "AFG", "Afghanistan", 1000.toDecimal(), mapOf(
+                    "deaths" to 10.toDecimal(),
+                    "cases" to 100.toDecimal()
+            )),
+            BurdenEstimate("Hib3", 1980, 30, "AGO", "Angola", 2000.toDecimal(), mapOf(
+                    "deaths" to 20.toDecimal(),
+                    "dalys" to 73.6.toDecimal()
+            ))
+    )
+
+    private val defaultProperties = CreateBurdenEstimateSet(
+            BurdenEstimateSetType(BurdenEstimateSetTypeCode.CENTRAL_UNKNOWN),
+            modelRunParameterSetId = null
+    )
+}
